Fix lints

+++ crates/protocol/src/poll/linux.rs
@@ -39,7 +39,7 @@ impl Poll {
             let mut ev: epoll_event = mem::zeroed();
 
             ev.events = interest.as_u32();
-            ev.u64 = token.0 as u64;
+            ev.u64 = token.0;
 
             if epoll_ctl(self.fd.as_raw_fd(), EPOLL_CTL_ADD, fd.as_raw_fd(), &mut ev) == -1 {
                 return Err(io::Error::last_os_error());
@@ -57,7 +57,7 @@ impl Poll {
             let mut ev: epoll_event = mem::zeroed();
 
             ev.events = interest.as_u32();
-            ev.u64 = token.0 as u64;
+            ev.u64 = token.0;
 
             if epoll_ctl(self.fd.as_raw_fd(), EPOLL_CTL_MOD, fd.as_raw_fd(), &mut ev) == -1 {
                 return Err(io::Error::last_os_error());
